Implement TimeScaleService

+++ Assets/Scripts/Core/Services/ITimeScaleService.cs
@@ -1,5 +1,5 @@
 public interface ITimeScaleService
 {
     void SetTimeScale(float timeScale);
-    int GetTimeScale();
+    float GetTimeScale();
 }
